+++ OleksandrMaslovskyj/src/interfaces/IBill.java
@@ -1,0 +1,16 @@
+package interfaces;
+
+
+import models.Product;
+
+public interface IBill {
+
+    Product addProduct(String name);
+
+    void closeBill();
+
+    double calculateAmountPrice();
+
+    void printBill();
+
+}
